+++ src/Sitecore.Pathfinder.Core/Msg.cs
@@ -67,6 +67,7 @@ namespace Sitecore.Pathfinder
         public const int C1059 = C + 59;
         public const int C1060 = C + 60;
         public const int C1061 = C + 61;
+        public const int C1062 = C + 62;
 
         public const int D = 4000;
         public const int D1000 = D;