+++ src/Sitecore.Pathfinder.Server/Extensions/DatabaseExtensions.cs
@@ -13,8 +13,8 @@ namespace Sitecore.Pathfinder.Extensions
 {
     public static class DatabaseExtensions
     {
-        [Diagnostics.NotNull,ItemNotNull]
-        public static IEnumerable<Item> GetItemsByTemplate([Diagnostics.NotNull] this Database database, [Diagnostics.NotNull, ItemNotNull]  params ID[] templateId)
+        [Diagnostics.NotNull, ItemNotNull]
+        public static IEnumerable<Item> GetItemsByTemplate([Diagnostics.NotNull] this Data.Database database, [Diagnostics.NotNull, ItemNotNull]  params Data.ID[] templateId)
         {
             var indexName = "sitecore_" + database.Name.ToLowerInvariant() + "_index";
 
@@ -27,8 +27,8 @@ namespace Sitecore.Pathfinder.Extensions
             }
         }
 
-        [ItemNotNull,Diagnostics.NotNull]
-        public static IEnumerable<Item> Query([Diagnostics.NotNull] this Database database, [Diagnostics.NotNull] string queryText)
+        [ItemNotNull, Diagnostics.NotNull]
+        public static IEnumerable<Item> Query([Diagnostics.NotNull] this Data.Database database, [Diagnostics.NotNull] string queryText)
         {
             var query = new Query(queryText)
             {