+++ src/Sitecore.Pathfinder.Server/WebApi/InstallProject.cs
@@ -6,7 +6,6 @@ using System.IO;
 using System.Net;
 using System.Text;
 using System.Web.Mvc;
-using Sitecore.Pathfinder.Emitters;
 using Sitecore.Pathfinder.Emitting;
 using Sitecore.Pathfinder.Extensions;